+++ static/app/views/monitors/components/overviewTimeline/index.tsx
@@ -17,7 +17,7 @@ import {CheckInTimeline} from 'sentry/views/monitors/components/checkInTimeline'
 import {
   GridLineOverlay,
   GridLineTimeLabels,
-} from 'sentry/views/monitors/components/overviewTimeline/timelineScrubber';
+} from 'sentry/views/monitors/components/overviewTimeline/gridLines';
 
 import {Monitor} from '../../types';
 import {scheduleAsText} from '../../utils';
